+++ activerecord/lib/active_record/connection_adapters/mysql/type_metadata.rb
@@ -2,13 +2,12 @@ module ActiveRecord
   module ConnectionAdapters
     module MySQL
       class TypeMetadata < DelegateClass(SqlTypeMetadata) # :nodoc:
-        attr_reader :extra, :strict
+        attr_reader :extra
 
-        def initialize(type_metadata, extra: "", strict: false)
+        def initialize(type_metadata, extra: "")
           super(type_metadata)
           @type_metadata = type_metadata
           @extra = extra
-          @strict = strict
         end
 
         def ==(other)
@@ -24,7 +23,7 @@ module ActiveRecord
         protected
 
           def attributes_for_hash
-            [self.class, @type_metadata, extra, strict]
+            [self.class, @type_metadata, extra]
           end
       end
     end